Jorge Torres Code Armada
This is the code I have so far. When the boss appears the models change
with the modified sprite sheet. I can change the dimensions of the
sprite sheet without issues, but still having issues drawing any type of
image other than the one specified. I removed the uploading code since
it only made the program crash.