Jorge Torres scrolling
I added a scene array that will track the backgrounds into the global
variables.
I also added some new methods to levelEntityClass which will set a
default scrolling animation, and can be overwritten later. Added a space
picture that I plan to use to pan through. The earth level entity is for
a future level creation idea. (nothing to do with the panning for now)