placing background into scenes instead of sprites
what i am trying to do here is use a scenes array that will handle the
animation and the updating of the backgrounds. rather than have the
sprite array handle everything, i would prefer this new array to be used
in order to add images to it later and be able to create the looping
effect better.