Fixed alien death animation
I just used MAXHEALTH to solve whether the alien is a basic one or
mothership, and change sourceX appropriately. Just a temporary fix until
we rework the tilesheet.
I also added placeholder sprites for the powerups that spawn. I started
work on a repair powerup, but since player health isn't implemented yet
I haven't completed it and it doesn't spawn.