Mayhem - Reorganization
The levels data (each level requires 2 arrays) is now in Levels.js
Changed the format of spriteObject to the way we are familiar.
Monster is now a child class of spriteObject, and changeDirection &
findClosestDirection are defined within that class.
Player is another child class. Its update function also contains the
camera scrolling.
Changed some of the code for the minimap.