Mayhem - Added timer, bomb (not yet implemented)
Imported the timer from the other maze game in chapter 10, and use it as
a loss condition.
Created an inventory. At this point the only real function is counting
stars collected, and bombs (which don't exist yet), but it can be
expanded on for different items.
Changed the arrays: Walls, boxes, monsters, stars, and player are all
sprite type objects. Floors are not. That way, during player.update, the
loop is less than half as big as it previously was.

Two issues, both with the inventory canvas.
- Lines 313 vs 314. I'm trying to do clearRect for inventory, but I
can't use inventory.width and inventory.height the way I use
minimap.width and minimap.height above, I have to hard-code the
dimensions for some reason
-Lines 412-414 vs 419-421. Trying to display numbers as counters of
items in the player's inventory, but my InventoryMessage object isn't
being recognized despite having created it the same way as all of the
other message objects. Putting the values in directly works fine, but
making references to attributes of inventoryMessage doesn't.