Mayhem - Minimap, camera changes. Pathfinding fix
Added a new second level, which is 8x8 instead of standard 16x16.
Changed camera control to work properly with maps small enough to fit
entirely within the camera.
Changed minimap to build off of variables, to scale with varying map
sizes.
Fixed a monster pathing issue, if it was spawned in a straight
up-and-down or straight left-and-right hallway, it would never start to
move.
Timer count accounts for negative numbers.
Bomb placement based off of player's centerX and centerY (instead of
just X and Y)