Mayhem - Improved inventory display
Lives are now displayed in the inventory.  All inventory content still
gets erased during level transitions, so I still have a global variable
named "lives", and the inventory's counter is dependent on that.
As of now, the blue stars do nothing, I don't even have any in the game.
I just used them in the inventory so I could figure out the workings of
columns there., and left them in there to show how it looks.