Hedgehog - Spritesheet code change
Objects now have an attribute, this.sheet, which will be "image" for
most sprites, except the player which will be "astro".
This way, for the update and the render, player can still be part of the
sprites array and be part of the same loops instead of having to do
separate calls for it.