Hedgehog - minor changes
On the spritesheet, the area surrounding the game message box (where
"You won"/"you lost" displays) are now transparent.
Fixed a monster AI issue, in which they would still try to see what was
beneath them -- even if they were on the bottom of the map.
Got rid of my 2 insane maps and replaced them with more reasonable
versions of a similar idea.
Added in more boxes on all the maps to make some jumps easier or less
punishing for a missed jump.