Crab - Some now shoot fireballs
There's an attribute in the crab class, this.shoots, which is true or
false. Currently I have it randomized, so half will be true and half
will be false.
There's another attribute, this.counter, which is initialized to a
random value between 1 and 15 seconds.
In the update function, for those that do shoot, every interval of 15
seconds, they'll shoot a fireball in the same direction as they're
walking.
They kill the player upon hitting them, and disappear if they hit a box
or go outside the gameworld.