Armada - Fleshed out powerup options
For each of the powerup choices, there is a further choice of whether
they spawn based on time or score (Future plans to give different
options, and frequency).

Player can also choose the skin of their ship.
The different shooting types will be in the options, rather than being
tied to the ship type (The choices are shown but function not yet
implemented)