Reset buttons changed - PARTIALLY functional
There is now a Settings class, which has an object that determines all
of the settings' values. At the end of Exploration.js is where the
options all get set based on those attributes.
PARTIALLY functional - Both buttons do the same thing. Next step is to
get the values to save to one of the Settings objects.